feat(forward+postprocess): restart writer, shallow vertex recording, spatial tiling
Forward:
- Add RestartWriter/RestartReader for full-volume restart state (latest-only)
- Add RecordingMap struct for shallow mesh-vertex strain extraction
- Change RecordWriter to write vertex_ids + strain at recording-map vertices only
- Add --resume flag support to gf_solver
- Read new config fields: record_depth_max_m/actual_m, green_tile_size_m,
  restart_dt_s, restart_stride
- Read /recording/ group from partition files
- Solver writes restart at restart_stride, resumes from restart file

Postprocess:
- Rewrite reader for new record format (vertex_ids, no NGLL dims)
- Add ConfigReader for config.h5 simulation attrs
- Rewrite writer with spatial tiling (green_tile_size_m bins)
- New tile schema: tile_x{i}_y{j}.h5 with vertex_ids, greens_tensor
- CLI takes config.h5 as argument, removes --tile-elems
- Tests updated: 15 postprocess + 102 general = 117 passing

+++ preprocess/model_writer.py
@@ -149,7 +149,7 @@ def _write_partition_files(
                     _write_dataset(rec_grp, "source_element_local_index",
                                    np.array(per_rank_rec["source_element_local_index"], dtype=np.int32), dtype="int32")
                     _write_dataset(rec_grp, "source_corner_index",
-                                   np.array(per_rank_rec["source_corner_index"], dtype=np.int8), dtype="int8")
+                                   np.array(per_rank_rec["source_corner_index"], dtype=np.int32), dtype="int32")
